fix(core): reject ZIP overlap, preserve full ordinal domain, Unicode-fold duplicate names, check required files before parsing
Addresses Task 2 review findings 2-5:
- archive.rs: call zip::ZipArchive::has_overlapping_files immediately after
  opening; overlap or the underlying read error maps to invalid_archive.
- model.rs/normalize.rs: ViewerLevel.ordinal widened to f64 so fractional and
  out-of-i32-range finite ordinals are preserved exactly (matching the
  browser's Number.isFinite(ordinalRaw) ? ordinalRaw : 0 contract) instead of
  being silently coerced to 0; level sort uses f64::total_cmp.
- archive.rs: duplicate-filename and required-file presence checks now use
  Unicode-aware str::to_lowercase (matching JS String.prototype.toLowerCase)
  instead of ASCII-only lowercasing.
- archive.rs: required-file presence is now validated before any entry
  content is parsed, so a missing required file is reported even when another
  entry has malformed JSON.

Finding 1 (manifest version acceptance) was reviewed and left unchanged: the
retained src/imdf/imdf.worker.ts regex and its existing pre-release-acceptance
test are the authoritative browser contract this task ports verbatim.

+++ core/crates/kiriko-model/src/model.rs
@@ -192,11 +192,14 @@ pub struct ImdfManifest {
     pub rest: Object,
 }
 
-/// Aggregated level row.
+/// Aggregated level row. `ordinal` preserves the full finite IEEE-754 domain
+/// the browser accepts (fractional, negative, or beyond `i32` range) rather
+/// than truncating to an integer; only non-finite/non-numeric values default
+/// to `0.0`. Matches `normalizeVenue.ts`'s `ordinalRaw` handling exactly.
 #[derive(Debug, Clone, PartialEq)]
 pub struct ViewerLevel {
     pub id: String,
-    pub ordinal: i32,
+    pub ordinal: f64,
     pub label: BTreeMap<String, String>,
     pub short_name: BTreeMap<String, String>,
 }
